Revert "classes fixes"
This reverts commit fba0185986b7e96f0b12124363f3f595c8c71cea.
