+++ Mindbox/Model/Request/ProductListRequest.swift
@@ -9,22 +9,12 @@ open class ProductListRequest: Encodable {
 
     public init(
         product: ProductRequest? = nil,
-        count: Decimal? = nil,
-        pricePerItem: Decimal? = nil,
-        priceOfLine: Decimal? = nil
-    ) {
-        self.product = product
-        self.count = count
-        self.pricePerItem = pricePerItem
-        self.priceOfLine = priceOfLine
-    }
-
-    public init(
         productGroup: ProductGroupRequest? = nil,
         count: Decimal? = nil,
         pricePerItem: Decimal? = nil,
         priceOfLine: Decimal? = nil
     ) {
+        self.product = product
         self.productGroup = productGroup
         self.count = count
         self.pricePerItem = pricePerItem
